fixup! Add style prefix function to support older Safari
